LPS-55100 Move to com.liferay.cobertura.agent
