+++ modules/test/cobertura-agent/src/com/liferay/cobertura/agent/ContextAwareClassWriter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.cobertura.instrument;
+package com.liferay.cobertura.agent;
 
 import java.io.IOException;
 import java.io.InputStream;
